+++ package.json
@@ -11,7 +11,7 @@
   "dependencies": {
     "cors": "^2.8.5",
     "fuse.js": "^6.6.2",
-    "next": "12.1.6",
+    "next": "^12.1.7-canary.45",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
